requirements.txt
kaggle~=1.7.4.5
matplotlib~=3.10.5
nltk~=3.9.1
numpy~=2.3.2
pandas~=2.3.1
Pillow~=11.3.0
PyYAML~=6.0.2
torch~=2.8.0
torchvision~=0.23.0
tqdm~=4.67.1

